created registerController and refactored code
