+++ server/routes/register.js
@@ -1,8 +1,7 @@
 const express = require("express");
 const router = express.Router();
-const bcrypt = require("bcryptjs");
-const User = require("../models/User");
-const { check, validationResult } = require("express-validator");
+const registerController = require("../controllers/registerController");
+const { check } = require("express-validator");
 
 router.post(
   "/",
@@ -16,37 +15,7 @@ router.post(
       .withMessage("not a valid email")
       .normalizeEmail()
   ],
-  async (req, res) => {
-    const errors = validationResult(req);
-    //check validators
-    if (!errors.isEmpty()) {
-      console.log(errors);
-      return res
-        .status(400)
-        .send({ msg: "Invalid submission", errors: errors });
-    }
-    const email = req.body.email;
-    const password = req.body.password;
-    try {
-      //hash password and save to mongodb
-      const hashedpwd = await bcrypt.hash(password, 10);
-      const user = new User({
-        email: email,
-        password: hashedpwd
-      });
-      const newUser = await user.save();
-      return res.status(200).send(newUser);
-    } catch (err) {
-      console.log(err);
-      if (err.code == 11000) {
-        // user already exists
-        return res.status(500).send({
-          msg: "User already registered"
-        });
-      }
-      return res.status(400).send(err);
-    }
-  }
+  registerController.registerUser
 );
 
 module.exports = router;
